Reflow by adding div.inners
After messing around a lot with the idea of using margins (which won't
paint background-color) or something like that, I instead decided that
I was going to try this.  One way to step back from this approach is
to simply calc() the padding-left and padding-right based on vw, which
worked according to my testing, but this is seemingly more elegant?

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>